+++ ReduxCore/inc/fields/media/field_media.php
@@ -66,6 +66,10 @@ if ( ! class_exists( 'ReduxFramework_media' ) ) {
                 'width'     => '',
                 'height'    => '',
                 'thumbnail' => '',
+                'title'     => '',
+                'caption'   => '',
+                'alt'       => '',
+                'description' => '',
             );
 
             $this->value = wp_parse_args( $this->value, $defaults );
@@ -156,10 +160,14 @@ if ( ! class_exists( 'ReduxFramework_media' ) ) {
             echo '<input placeholder="' . $placeholder . '" type="text" class="' . $hide . 'upload large-text ' . $this->field['class'] . '" name="' . $this->field['name'] . $this->field['name_suffix'] . '[url]" id="' . $this->parent->args['opt_name'] . '[' . $this->field['id'] . '][url]" value="' . $this->value['url'] . '"' . $readOnly . '/>';
             echo '<input type="hidden" class="data" data-mode="' . $this->field['mode'] . '" />';
             echo '<input type="hidden" class="library-filter" data-lib-filter="' . $libFilter . '" />';
-            echo '<input type="hidden" class="upload-id ' . $this->field['class'] . '" name="' . $this->field['name'] . $this->field['name_suffix'] . '[id]" id="' . $this->parent->args['opt_name'] . '[' . $this->field['id'] . '][id]" value="' . $this->value['id'] . '" />';
-            echo '<input type="hidden" class="upload-height" name="' . $this->field['name'] . $this->field['name_suffix'] . '[height]" id="' . $this->parent->args['opt_name'] . '[' . $this->field['id'] . '][height]" value="' . $this->value['height'] . '" />';
-            echo '<input type="hidden" class="upload-width" name="' . $this->field['name'] . $this->field['name_suffix'] . '[width]" id="' . $this->parent->args['opt_name'] . '[' . $this->field['id'] . '][width]" value="' . $this->value['width'] . '" />';
-            echo '<input type="hidden" class="upload-thumbnail" name="' . $this->field['name'] . $this->field['name_suffix'] . '[thumbnail]" id="' . $this->parent->args['opt_name'] . '[' . $this->field['id'] . '][thumbnail]" value="' . $this->value['thumbnail'] . '" />';
+            echo '<input type="hidden" class="upload-id ' . $this->field['class'] . '" name="' . $this->field['name'] . $this->field['name_suffix'] . '[id]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_id" value="' . $this->value['id'] . '" />';
+            echo '<input type="hidden" class="upload-height" name="' . $this->field['name'] . $this->field['name_suffix'] . '[height]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_height" value="' . $this->value['height'] . '" />';
+            echo '<input type="hidden" class="upload-width" name="' . $this->field['name'] . $this->field['name_suffix'] . '[width]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_width" value="' . $this->value['width'] . '" />';
+            echo '<input type="hidden" class="upload-thumbnail" name="' . $this->field['name'] . $this->field['name_suffix'] . '[thumbnail]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_thumbnail" value="' . $this->value['thumbnail'] . '" />';
+            echo '<input type="hidden" class="upload-title" name="' . $this->field['name'] . $this->field['name_suffix'] . '[title]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_title" value="' . $this->value['title'] . '" />';
+            echo '<input type="hidden" class="upload-caption" name="' . $this->field['name'] . $this->field['name_suffix'] . '[caption]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_caption" value="' . $this->value['caption'] . '" />';
+            echo '<input type="hidden" class="upload-alt" name="' . $this->field['name'] . $this->field['name_suffix'] . '[alt]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_alt" value="' . $this->value['alt'] . '" />';
+            echo '<input type="hidden" class="upload-description" name="' . $this->field['name'] . $this->field['name_suffix'] . '[description]" id="' . $this->parent->args['opt_name'] . '_' . $this->field['id'] . '_description" value="' . $this->value['description'] . '" />';
 
             //Preview
             $hide = '';